PMM-7259 move EmptyBlock inside Tabs
