+++ public/app/features/integrated-alerting/components/IntegratedAlertingContent/IntegratedAlertingContent.styles.ts
@@ -2,6 +2,10 @@ import { css } from 'emotion';
 import { GrafanaTheme } from '@grafana/data';
 
 export const getStyles = ({ colors }: GrafanaTheme) => ({
+  emptyBlock: css`
+    margin-top: 40px;
+    padding: 8px;
+  `,
   link: css`
     color: ${colors.linkExternal};
     &:hover {
