--- public/app/features/integrated-alerting/components/IntegratedAlertingContent/IntegratedAlertingTabs/index.ts
@@ -1,1 +1,0 @@
-export * from './IntegratedAlertingTabs';
